Fixes from review.
