+++ sdk/metrics-testing/src/main/java/io/opentelemetry/sdk/testing/assertj/metrics/LongPointDataAssert.java
@@ -10,7 +10,7 @@ import org.assertj.core.api.Assertions;
 
 /** Test assertions for {@link LongPointData}. */
 public class LongPointDataAssert
-    extends AbstractSampledPointDataAssert<LongPointDataAssert, LongPointData> {
+    extends AbstractPointDataAssert<LongPointDataAssert, LongPointData> {
 
   protected LongPointDataAssert(LongPointData actual) {
     super(actual, LongPointDataAssert.class);
