+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/data/DoubleSummaryPointData.java
@@ -7,6 +7,7 @@ package io.opentelemetry.sdk.metrics.data;
 
 import com.google.auto.value.AutoValue;
 import io.opentelemetry.api.common.Attributes;
+import java.util.Collections;
 import java.util.List;
 import javax.annotation.concurrent.Immutable;
 
@@ -17,6 +18,17 @@ import javax.annotation.concurrent.Immutable;
 @Immutable
 @AutoValue
 public abstract class DoubleSummaryPointData implements PointData {
+  /**
+   * Creates a {@link DoubleSummaryPointData}.
+   *
+   * @param startEpochNanos (optional) The starting time for the period where this point was
+   *     sampled.
+   * @param epochNanos The ending time for the period when this value was sampled.
+   * @param attributes The set of attributes associated with this point.
+   * @param count The number of measurements being sumarized.
+   * @param sum The sum of measuremnts being sumarized.
+   * @param percentileValues Calculations of percentile values from measurements.
+   */
   public static DoubleSummaryPointData create(
       long startEpochNanos,
       long epochNanos,
@@ -25,7 +37,13 @@ public abstract class DoubleSummaryPointData implements PointData {
       double sum,
       List<ValueAtPercentile> percentileValues) {
     return new AutoValue_DoubleSummaryPointData(
-        startEpochNanos, epochNanos, attributes, count, sum, percentileValues);
+        startEpochNanos,
+        epochNanos,
+        attributes,
+        Collections.emptyList(),
+        count,
+        sum,
+        percentileValues);
   }
 
   DoubleSummaryPointData() {}
