+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/data/LongPointData.java
@@ -19,7 +19,7 @@ import javax.annotation.concurrent.Immutable;
  */
 @Immutable
 @AutoValue
-public abstract class LongPointData implements SampledPointData {
+public abstract class LongPointData implements PointData {
 
   LongPointData() {}
